refactor(update-apis-reference): Moved use-axios-swr from hooks to http directory, and updated all the apis to reflect the change
