+++ lib/api/contact-details/v1/service.ts
@@ -2,9 +2,9 @@ import { config } from "@mtfh/common/lib/config";
 import {
   AxiosSWRConfiguration,
   AxiosSWRResponse,
+  axiosInstance,
   useAxiosSWR,
-} from "@mtfh/common/lib/hooks";
-import { axiosInstance } from "@mtfh/common/lib/http";
+} from "@mtfh/common/lib/http";
 import { removeWhitespace } from "@mtfh/common/lib/utils";
 import {
   ContactDetail,
