+++ lib/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from "./use-axios-swr";
 export * from "./use-breakpoint";
 export * from "./use-error-codes";
 export * from "./use-feature-toggle";
